clin-666: refactor fonctio nanuq
